fix float

+++ 0x10-variadic_functions/3-print_all.c
@@ -48,7 +48,7 @@ void print_str(va_list args)
  */
 void print_float(va_list args)
 {
-	printf("%f", va_arg(args, double));
+	printf("%f", (float) va_arg(args, double));
 }
 
 /**
